refactor: hooks + auth helpers added | mz
